+++ examples/v2_get_first_block_epoch.rs
@@ -1,0 +1,34 @@
+//! Test the `GetFirstBlockEpoch` endpoint.
+use anyhow::Context;
+use clap::AppSettings;
+use concordium_rust_sdk::v2;
+use structopt::StructOpt;
+
+#[derive(StructOpt)]
+struct App {
+    #[structopt(
+        long = "node",
+        help = "GRPC interface of the node.",
+        default_value = "http://localhost:20000"
+    )]
+    endpoint: v2::Endpoint,
+
+    #[structopt(long = "epoch", help = "Epoch identifier", default_value = "%5,0")]
+    epoch: v2::EpochIdentifier,
+}
+
+#[tokio::main(flavor = "multi_thread")]
+async fn main() -> anyhow::Result<()> {
+    let app = {
+        let app = App::clap().global_setting(AppSettings::ColoredHelp);
+        let matches = app.get_matches();
+        App::from_clap(&matches)
+    };
+
+    let mut client = v2::Client::new(app.endpoint)
+        .await
+        .context("Cannot connect.")?;
+    let fb = client.get_first_block_epoch(app.epoch).await?;
+    println!("{:?}", fb);
+    Ok(())
+}