+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 ## Unreleased changes
-- Add a `commission_rates` field to `CurrentPaydayBakerPoolStatus` which yields the commission rates 
-  of the baker for the reward period (requires a node with version at least 6.1).
+- Add a `commission_rates` field to `CurrentPaydayBakerPoolStatus` which yields the commission rates
+  of the baker for the reward period. Requires a node version at least 6.1.
+- Add support for `GetWinningBakersEpoch`. Requires a node version at least 6.1.
+- Add Support for `GetFirstBlockEpoch`. Requires a node version at least 6.1.
+- Add support for `GetBakersRewardPeriod` endpoint. Requires a node version at least 6.1.
+- Add Support for `GetBakerEarliestWinTime` endpoint. Requires a node version at least 6.1.
 - Add support for `GetBlockCertificates`. Requires a node version at least 6.1.
+
+## 3.0.1
+
+- Update `concordium_base` dependency to 3.0.1.
 
 ## 3.0.0